Separate UserPresence component from DocumentHeader component
